refactor - 메인 View 패키지 이동
