+++ presentation/src/main/java/com/preonboarding/customkeyboard/presentation/main/MainViewModel.kt
@@ -1,4 +1,4 @@
-package com.preonboarding.customkeyboard.presentation
+package com.preonboarding.customkeyboard.presentation.main
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
